We want *all* the reads in the output file
With the `-0` option, the output file ("interleaved") is empty for PacBio
because all the reads go to the "other" file. Whereas paired-reads all go to
the "interleaved" file and none to the "other" file.

The simplest is to not use the `-0` option. Then, `samtools fasta` simply sends
all the reads to the standard output.